Move metadata routines back into client part of library
