+++ func_adl/__init__.py
@@ -5,6 +5,7 @@ from .event_dataset import EventDataset, find_EventDataset
 
 # Extra LINQ-like functions
 from .functions import Range
+from .meta_data import extract_metadata  # NOQA
 
 # Main streaming object
 from .object_stream import ObjectStream
